Move components into individual files
