+++ src/components/PeopleTable/PeopleTableHeader.jsx
@@ -1,0 +1,27 @@
+/* eslint-disable react/jsx-key */
+import React from 'react';
+import PropTypes from 'prop-types';
+import { Th, Tr, Flex } from '@chakra-ui/react';
+import styles from './PeopleTable.module.css';
+
+const PeopleTableHeader = ({ headerGroups, loading }) => {
+  return headerGroups.map(headerGroup => (
+    <Tr className={styles['table-head']} {...headerGroup.getHeaderGroupProps()}>
+      {headerGroup.headers.map(column => (
+        <Th {...column.getHeaderProps()} color="white" bgColor="ochGrey">
+          <Flex alignItems="center" textTransform="none">
+            {loading ? <>&nbsp;</> : column.render('Header')}
+          </Flex>
+        </Th>
+      ))}
+    </Tr>
+  ));
+};
+
+PeopleTableHeader.propTypes = {
+  // eslint-disable-next-line react/forbid-prop-types
+  headerGroups: PropTypes.object.isRequired,
+  loading: PropTypes.bool.isRequired,
+};
+
+export default PeopleTableHeader;
